# frontend/edilgiulianiUI/src/app/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Openapi specs",
    "description": "OpenApi doc for spring backend edilgiuliani",
    "contact": {
      "name": "Adriano",
      "email": "glndrn02m02i872f@studenti.unical.it"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/booking/new": {
      "post": {
        "tags": [
          "booking-controller"
        ],
        "operationId": "prenota",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrenotazioneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/PrenotazioneResponse"
                }
              }
            }
          }
        }
      }
    },
    "/booking/deleteP": {
      "post": {
        "tags": [
          "booking-controller"
        ],
        "operationId": "deleteP",
        "parameters": [
          {
            "name": "idp",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/booking/allhours": {
      "get": {
        "tags": [
          "booking-controller"
        ],
        "operationId": "getBookedHours",
        "parameters": [
          {
            "name": "idmezzo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "day",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IsoInterval"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/booking/alldays": {
      "get": {
        "tags": [
          "booking-controller"
        ],
        "operationId": "getBookedDays",
        "parameters": [
          {
            "name": "idmezzo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Mezzo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nome": {
            "type": "string"
          },
          "descrizione": {
            "type": "string"
          },
          "prezzo_hour": {
            "type": "number",
            "format": "float"
          },
          "minPrice": {
            "type": "number",
            "format": "float"
          },
          "dailyPrice": {
            "type": "number",
            "format": "float"
          },
          "available": {
            "type": "boolean"
          }
        }
      },
...
      "PrenotazioneResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nome": {
            "type": "string"
          },
          "indirizzo": {
            "type": "string"
          },
          "prezzoStimato": {
            "type": "number",
            "format": "float"
          },
          "dataCreazione": {
            "type": "string",
            "format": "date-time"
          },
          "descrizione": {
            "type": "string"
          },
          "cap": {
            "type": "string"
          },
          "subB":{
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubBooking"
            }
          }
        }
      },
...
      "IsoInterval": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string"
          },
          "end": {
            "type": "string"
          }
        }
      },
      "PrenotazioneRequest": {
        "required": [
          "cap",
          "descrizione",
          "indirizzo",
          "nome",
          "subB"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nome": {
            "type": "string"
          },
          "indirizzo": {
            "type": "string"
          },
          "cap": {
            "type": "string"
          },
          "subB": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubBooking"
            }
          },
          "descrizione": {
            "type": "string"
          }
        }
      },
      "SubBooking": {
        "type": "object",
        "properties": {
          "mezzo": {
            "$ref": "#/components/schemas/Mezzo"
          },
          "intervals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IsoInterval"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "oauth2",
        "description": "keycloak jwt",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "flows": {
          "clientCredentials": {
            "authorizationUrl": "http://localhost:8081/realms/edilgiuliani2-realm/protocol/openid-connect/auth",
            "scopes": {}
          }
        }
      }
    }
  }
}